HT_13: added realized the home tasks (beta)

+++ HT_13/task_2/service/person_book_service.py
@@ -1,4 +1,4 @@
-from HT_13.task_2.model.model import Book, Person
+from HT_13.task_2.model.model import Book, Person, CustomException
 from HT_13.task_2.repository.person_book_repository import PersonBookRepository
 
 
@@ -17,7 +17,7 @@ class PersonBookService:
                     persons_dict[person] = [book]
         return persons_dict
 
-    def get_person_with_books(self, person:Person) -> dict[Person:list[Book]]:
+    def get_person_with_books(self, person:Person) -> list[Book]:
         return self.__repository.get_by_person_id(person.id)
 
     def get_books_with_persons(self) -> dict[Book:list[Person]]:
@@ -32,7 +32,12 @@ class PersonBookService:
         return books_dict
 
     def save(self, person: Person, book: Book):
+        if any(i.id == book.id for i in self.get_person_with_books(person)):
+            raise CustomException("person have this book")
         self.__repository.insert(person.id, book.id)
 
     def delete(self, person: Person, book: Book):
+        if not any(i.id == book.id for i in self.get_person_with_books(person)):
+            raise CustomException("person have not this book")
+
         self.__repository.delete(person.id, book.id)
